pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ec349-project"
version = "0.1.0"
description = "A project for EC349"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0",
    "transformers>=4.30",
    "scikit-learn>=1.3",
    "pandas>=1.5",
    "numpy>=1.23",
    "pytorch-lightning>=2.0",
    "pyarrow>=12.0",
    "python-dotenv>=1.0",
    "tqdm>=4.65",
    "pytest>=7.4",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]